fix getPath tests on windows
